Former-commit-id: 2275fb560e10192faecc258ad920be3f021c3f02

+++ deployment/utilities/iam-deployment.js
@@ -1,10 +1,8 @@
 'use strict';
 const _ = require('underscore');
 const du = global.SixCRM.routes.include('lib', 'debug-utilities.js');
-const eu = global.SixCRM.routes.include('lib', 'error-utilities.js');
 const fileutilities = global.SixCRM.routes.include('lib', 'file-utilities.js');
 const arrayutilities = global.SixCRM.routes.include('lib', 'array-utilities.js');
-const objectutilities = global.SixCRM.routes.include('lib', 'object-utilities.js');
 const parserutilities = global.SixCRM.routes.include('lib', 'parser-utilities.js');
 const timestamp = global.SixCRM.routes.include('lib', 'timestamp.js');
 const AWSDeploymentUtilities = global.SixCRM.routes.include('deployment', 'utilities/aws-deployment-utilities.js');
@@ -147,7 +145,7 @@ class IAMDeployment extends AWSDeploymentUtilities {
 
       create_parameters.AssumeRolePolicyDocument = parserutilities.parse(create_parameters.AssumeRolePolicyDocument, {aws_account_id: global.SixCRM.configuration.site_config.aws.account});
 
-      return this.iamutilities.createRole(create_parameters).then((result) => {
+      return this.iamutilities.createRole(create_parameters).then(() => {
 
         return timestamp.delay(3000)().then(() => this.addPoliciesAndPermissions(role_definition));
 
@@ -238,7 +236,12 @@ class IAMDeployment extends AWSDeploymentUtilities {
           let role_definitions = this.acquireRoleFile(role_file_name);
 
           let deploy_role_promises = arrayutilities.map(role_definitions, (role_definition) => {
-            return this.destroyRole(role_definition);
+
+						//Technical Debt: this can be way cleaner, can remove role from instance profile then delete
+						if(!role_definition.RoleName === 'DataPipelineDefaultResourceRole'){
+							return this.destroyRole(role_definition);
+						}
+
           });
 
           return Promise.all(deploy_role_promises);
@@ -283,7 +286,26 @@ class IAMDeployment extends AWSDeploymentUtilities {
 
       return this.iamutilities.deleteRole(delete_parameters);
 
-    }
+		}
+
+		//Technical Debt: need place intance profile settings into a config file
+		deployInstanceProfiles(){
+
+			du.debug('Deploy Instance Profile');
+
+			let parameters = {
+				InstanceProfileName : 'DataPipelineDefaultResourceRole'
+			}
+
+			return this.iamutilities.createInstanceProfile(parameters)
+				.then(() => {
+
+					parameters.RoleName = 'DataPipelineDefaultResourceRole'
+
+					return this.iamutilities.addRoleToInstanceProfile(parameters);
+				});
+
+		}
 
 }
